fix screenshot and zoom

+++ scripts/dot-config/scripts/sway_zoom.sh
@@ -1,4 +1,7 @@
-TMP=$(mktemp)
-grim -g "$(slurp)" - >"$TMP"
-imv -u nearest_neighbour "$TMP"
-rm "$TMP"
+XY="$(slurp)"
+if [ $? -eq 0 ]; then
+    TMP=$(mktemp)
+    grim -g "$XY" - >"$TMP"
+    imv -u nearest_neighbour "$TMP"
+    rm "$TMP"
+fi
